requirements.txt
tavily-python>=0.7.0      # Tavily搜索API客户端
python-dotenv>=0.19.0     # 环境变量管理